refactor: flyout components
