+++ src/components/Flyout.tsx
@@ -1,26 +1,26 @@
-import { useSelector } from "react-redux";
-import { RootState } from "../store";
-import FlyoutText from "./FlyoutText";
-import FlyoutDeselectButton from "./FlyoutDeselectButton";
-import FlyoutDownloadButton from "./FlyoutDownloadButton";
 import styles from "./Flyout.module.css";
 
-export default function Flyout() {
-  const { selectedItems } = useSelector(
-    (state: RootState) => state.selectedItems,
-  );
-
-  if (selectedItems.length === 0) {
-    return null;
-  }
-
+export default function Flyout({
+  itemsCount,
+  downloadUrl,
+  onDeselectAll,
+}: {
+  itemsCount: number;
+  downloadUrl: string;
+  onDeselectAll: () => void;
+}) {
   return (
     <div className={styles.flyout}>
-      <FlyoutText length={selectedItems.length} />
-      <FlyoutDeselectButton />
-      <FlyoutDownloadButton selected={selectedItems}>
+      <p>{`${itemsCount} ${itemsCount === 1 ? "item" : "items"} selected`}</p>
+      <button onClick={onDeselectAll}>Deselect all</button>
+      <a
+        className={styles.downloadButton}
+        href={downloadUrl}
+        download={`${itemsCount}_characters.csv`}
+        role="button"
+      >
         Download
-      </FlyoutDownloadButton>
+      </a>
     </div>
   );
 }
